Refactor stateless benchmark orchestration
