+++ benchmarks/schemas/held-out-manifest.schema.json
@@ -113,7 +113,7 @@
     "treatment": {
       "type": "object",
       "additionalProperties": false,
-      "required": ["id", "role", "jacobian_enabled", "image", "source_sha", "platform", "server_version", "policy_profile", "catalog_digest", "policy_digest"],
+      "required": ["id", "role", "jacobian_enabled", "image", "source_sha", "platform", "server_version", "catalog_digest"],
       "properties": {
         "id": {"const": "C2"},
         "role": {"const": "PRIMARY_TREATMENT"},
@@ -122,9 +122,7 @@
         "source_sha": {"type": "string", "pattern": "^[0-9a-f]{40}$"},
         "platform": {"type": "string", "pattern": "^[a-z0-9_]+/[a-z0-9_]+$"},
         "server_version": {"type": "string", "minLength": 1},
-        "policy_profile": {"type": "string", "minLength": 1},
-        "catalog_digest": {"$ref": "#/$defs/digest"},
-        "policy_digest": {"$ref": "#/$defs/digest"}
+        "catalog_digest": {"$ref": "#/$defs/digest"}
       }
     },
     "stage": {
